25.48W_tokentest/claude/claude/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pillow
python-multipart
pydantic
python-dotenv
tiktoken>=0.7.0
